+++ Xendit/M2Invoice/view/frontend/web/js/view/payment/method-renderer/alfamart.js
@@ -32,6 +32,10 @@ define(
                 return 'Alfamart';
             },
 
+            getMethodImage: function () {
+                return window.checkoutConfig.payment[this.item.method].image;
+            },
+
             getTest: function() {
                 return '1';
             },